typings: Split the DeviceTypeJson namespace types to a different file
Change-type: patch
